requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
cryptography==42.0.5
Flask==2.3.2
Flask-Migrate
flask-restx
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mypy
mysqlclient==2.2.4
pymysql==1.1.0
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.28
typeshed_client==2.5.1
typing_extensions==4.10.0